requirements.txt
numpy==1.21.6
pandas==1.3.3
streamlit==1.0.0
psycopg2-binary==2.9.1
mplsoccer==1.0.2
python-dotenv==0.19.0

